requirements.txt
inflection==1.4.0
azure.cognitiveservices.speech==1.9.0